web - introduce `web.api.ts`
