+++ src/vs/workbench/services/environment/common/environmentService.ts
@@ -6,9 +6,9 @@
 import { refineServiceDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { IWindowConfiguration } from 'vs/platform/windows/common/windows';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
-// TODO@layers
-// eslint-disable-next-line code-import-patterns
-import type { IWorkbenchConstructionOptions as IWorkbenchOptions } from 'vs/workbench/workbench.web.api';
+// TODO@bpasero layers
+// eslint-disable-next-line code-layering, code-import-patterns
+import type { IWorkbenchConstructionOptions as IWorkbenchOptions } from 'vs/workbench/browser/web.api';
 import { URI } from 'vs/base/common/uri';
 
 export const IWorkbenchEnvironmentService = refineServiceDecorator<IEnvironmentService, IWorkbenchEnvironmentService>(IEnvironmentService);
